refactor: Move map info
