+++ dialogs/MapInformation.h
